+++ src/main/java/domain/Score.java
@@ -1,33 +1,2 @@
-package domain;
-
-public class Score {
-    private int strike;
-    private int ball;
-
-    public Score() {
-    }
-
-    public void addStrike() {
-        strike += 1;
-    }
-
-    public void addBall() {
-        ball += 1;
-    }
-
-    public boolean isNothing() {
-        return strike == 0 && ball == 0;
-    }
-
-    public boolean isAllStrike() {
-        return strike == 3;
-    }
-
-    public int getStrike() {
-        return strike;
-    }
-
-    public int getBall() {
-        return ball;
-    }
+package domain;public class Score {
 }
